move counter outside Richtext component
